Implemented mongodb configs to infer URI and setup more complex databases
