+++ docker/docker-compose.yaml
@@ -1,14 +1,20 @@
 #  Raymond Kirk (Tunstill) Copyright (c) 2020
 #  Email: ray.tunstill@gmail.com
 
-# set environment variables using .env file located in this same directory
+# Not for direct use, see topic_store/docker/docker_compose_up_safe.sh
+
+# To dump the database once running:
+# $ docker exec ts_mongodb_server sh -c 'exec mongodump -d <database_name> --archive' > /some/host/path/all-collections.archive
 version: "3.7"
 
 services:
   mongodb:
-    container_name: mongodb_server
+    container_name: ts_mongodb_server
     image: "mongo:4.2"
     ports:
-      - "65530:27017" # ExposedOn:DockerPort
+      - ${MONGO_net_port}:${MONGO_net_port} # Bind the docker IP to the same local IP
     volumes:
-      - ${MONGO_DB_PATH}:/data/db
+      - ${TOPIC_STORE_ROOT}:${TOPIC_STORE_ROOT} # All files in the topic_store package available in the docker
+      - ${MONGO_DB_CONFIG_PATH}:${MONGO_DB_CONFIG_PATH} # Config available in the docker
+      - ${MONGO_storage_dbPath}:${MONGO_storage_dbPath} # Database folder mounted in the docker
+    command: --config "${MONGO_DB_CONFIG_PATH}"
